Use expandable buffer in Encoder
Avoids unnecessary buffer copying when creating large requests.
